units: allow conversions of quantities between commensurable units
Using the unit scaling factors, convert from one unit to another while
scaling the value to match so that the result is equivalent.

At this time, there is some inaccuracy introduced when moving between
metric and imperial measures, which may be addressed later.

+++ units/unit.go
@@ -17,7 +17,7 @@ type Unit struct {
 	//
 	// Scaling is only used for derived units. Units of base dimensions are
 	// just represented directly.
-	scale int
+	scale int64
 }
 
 var dimless = &Unit{Dimensionality{}, baseUnits{}, 0}
